python-workers/skipper-bot/requirements.txt
requests==2.31.0
supabase==2.3.4
python-dotenv==1.0.1
playwright==1.42.0
fake-useragent==1.5.1
redis==5.0.3
httpx==0.27.0
